feat: add Spanish Language (#31)
* * 🐳 chore(Dockerfile): add Dockerfile for sqlchat
This commit adds a Dockerfile for sqlchat. The Dockerfile is based on Alpine Linux version 3.17 and installs the necessary dependencies for the application to run. The Dockerfile also sets the metadata for the container and copies the application files into the container. The Dockerfile uses pnpm to install the application dependencies and sets the OPENAI_API_KEY and OPENAI_API_ENDPOINT arguments for the container. The CMD instruction specifies the command to run when the container starts, which is to run the application in development mode using pnpm.

* * 🌐 feat(LocaleSelector.tsx): add support for Spanish language
* 🌐 feat(LocaleSwitch.tsx): add support for switching to Spanish language
* 🎨 style(ThemeSwitch.tsx): refactor button JSX for readability
The LocaleSelector component now includes an option for Spanish language. The LocaleSwitch component now includes support for switching to Spanish language. The ThemeSwitch component has been refactored to improve readability of the button JSX.

* 🌐 feat(i18n.ts): add support for Spanish language
A new language, Spanish, has been added to the application. The translations for the new language have been added to the es.json file. The i18n configuration has been updated to include the new language.

* ✨ feat(setting.ts): add support for Spanish language
The Locale type now includes "es" for Spanish language support.

* * 🗑️ chore(Dockerfile): remove Dockerfile
The Dockerfile has been removed from the repository.

* * 🐛 fix(LocaleSwitch.tsx): fix indentation and add missing semicolon
* 🐛 fix(ThemeSwitch.tsx): remove unnecessary code and fix indentation
The LocaleSwitch component had an indentation issue and a missing semicolon. The ThemeSwitch component had unnecessary code that was removed, and the indentation was fixed.