* ✨ feat(README.md): add support for MSSQL database * ✨ feat(CreateConnectionModal.tsx): add support for MSSQL engine type * ✨ feat(EngineIcon.tsx): add support for MSSQL engine type icon * ✨ feat(connectors/index.ts): add support for MSSQL engine type connector MSSQL database is now supported by the application. The CreateConnectionModal component now has an option to select MSSQL as the engine type. The EngineIcon component now has an icon for MSSQL engine type. The connectors/index.ts file now has a connector for MSSQL engine type. The README.md file has been updated to reflect the addition of MSSQL support.
* 🎉 feat(mssql): add support for Microsoft SQL Server
This commit adds a new connector for Microsoft SQL Server. The connector supports the following features:
- Test connection
- Execute statement
- Get databases
- Get tables
- Get table structure

The connector uses the mssql package to connect to the database. The connector is created using a factory function that takes a Connection object as an argument and returns a Connector object.

* ✨ feat(connection.ts): add MSSQL engine and SSL options
The MSSQL engine has been added to the list of supported engines. Additionally, SSL options have been added to the Connection interface, including the ability to enable SSL and trust the server certificate.