chore: update on Spanish locale file (#73)
* 🎨 style(es.json): add new keys and values for connection and assistant
🚀 chore(es.json): add new banner messages
The connection object now has new keys and values for the database type, title, host, port, database name, username, and password. The assistant object now has a new key and value for creating a new bot. The banner object now has two new messages, one for warning non-select SQL statements and another for promoting the product launch on Product Hunt.

* 🎨 style(es.json): fix typo in banner.product-hunt message
The message in banner.product-hunt contained a typo. The word "de" was removed to make the message grammatically correct.

* 🌟 feat(es.json): add new translations for edit and execute
🐛 fix(es.json): update conversation title translation
🐛 fix(es.json): add new translations for OpenAI quota overflow
The new translations for "edit" and "execute" have been added to the language file. The translation for "conversation-title" has been updated to "title" for consistency with the application. The translations for "quota-overflow" and "use-my-key" have been added to handle OpenAI quota overflow.